Dodoi-Lab: changes types.

+++ src/ui/button.h
@@ -10,18 +10,18 @@
 
 typedef void (*button_onclick_callback_t)(void);
 typedef struct {
-    int x;
-    int y;
-    int w;
-    int h;
+    i32 x;
+    i32 y;
+    i32 w;
+    i32 h;
     bool visible;
-    uint16_t pad_button;
+    u8 pad_button;
     SDL_Texture* texture;
     button_onclick_callback_t onclick;
 } button_t;
 
-button_t button(int x, int y);
-button_t* button_new(int x, int y);
-void button_set_source(button_t* button, const uint8_t* data, size_t size);
+button_t button(i32 x, i32 y);
+button_t* button_new(i32 x, i32 y);
+void button_set_source(button_t* button, const u8* data, size_t size);
 void button_set_onclick(button_t* button, button_onclick_callback_t callback);
 void button_render(const button_t* button);
